requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
openai>=1.30.0
httpx>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
gradio>=4.31.0
openenv-core>=0.2.3
numpy>=1.24.0
fpdf2>=2.7.0
# Qiskit 0.46.3 uses legacy APIs (BasicAer, PauliSumOp, QuantumInstance).
# The protein folding module gracefully falls back to cached results if
# Qiskit is unavailable or the API surface has changed (Qiskit 1.x).
qiskit==0.46.3
# TRIBE v2 (Meta fMRI prediction model) is cloned locally in tribev2/.
# Its heavy dependencies (torch>=2.5.1, x_transformers, moviepy, etc.)
# are NOT listed here — the dashboard runs in "Sensory Simulation Mode"
# To run TRIBE v2 inference: pip install -e tribev2/
# Optional: pip install -e tribev2/ to enable TRIBE v2 inference
pennylane>=0.38.0
pennylane-qiskit
qiskit
rdkit
